requirements.txt
pytest
pyyaml==6.0.1
langchain_openai